+++ src/Models/Actions/ReorderFolder.php
@@ -7,8 +7,8 @@ namespace Hirtz\Media\Models\Actions;
 use Hirtz\Media\Models\Collections\FolderCollection;
 use Hirtz\Media\Models\Folder;
 use Hirtz\Skeleton\Models\Actions\ReorderActiveRecords;
+use Hirtz\Skeleton\I18n\Message;
 use Hirtz\Skeleton\Models\Trail;
-use Yii;
 
 class ReorderFolder extends ReorderActiveRecords
 {
@@ -26,7 +26,7 @@ class ReorderFolder extends ReorderActiveRecords
 
     protected function afterReorder(): void
     {
-        Trail::createOrderTrail(null, Yii::t('media', 'REORDER_FOLDER_FOLDER_ORDER_CHANGED'));
+        Trail::createOrderTrail(null, Message::make('media', 'REORDER_FOLDER_FOLDER_ORDER_CHANGED'));
         FolderCollection::invalidateCache();
 
         parent::afterReorder();