+++ src/Modules/Admin/Controllers/Traits/FolderControllerTrait.php
@@ -5,19 +5,14 @@ declare(strict_types=1);
 namespace Hirtz\Media\Modules\Admin\Controllers\Traits;
 
 use Hirtz\Media\Models\Folder;
-use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 
 trait FolderControllerTrait
 {
-    protected function findFolder(int $id, ?string $permissionName = null): Folder
+    protected function findFolder(int $id): Folder
     {
         if (!$folder = Folder::findOne($id)) {
             throw new NotFoundHttpException();
-        }
-
-        if ($permissionName && !$this->webuser->can($permissionName, ['folder' => $folder])) {
-            throw new ForbiddenHttpException();
         }
 
         return $folder;